use not connected in app and create useSocketEvents hook
